+++ src/base/eCP/ECPFarthestNeighbour.h
@@ -14,7 +14,7 @@ namespace exq {
     template<typename T, typename U, typename V>
     class ECPFarthestNeighbour {
     public:
-        ECPFarthestNeighbour(vector<double>& query, double bias, uint64_t k,
+        ECPFarthestNeighbour(vector<double>& query, double bias, uint32_t k,
                              ExqFunctions<ExqDescriptor<T,U,V>>*& functions);
 
         ~ECPFarthestNeighbour();
@@ -22,26 +22,26 @@ namespace exq {
         // Get info
         vector<double> getQuery() { return _query; }
 
-        uint64_t getK() { return _k; }
-        void setK(uint64_t k) { _k = k; }
+        uint32_t getK() { return _k; }
+        void setK(uint32_t k) { _k = k; }
 
         vector<double> getTopDistances();
 
-        vector<uint64_t> getTopIds();
+        vector<uint32_t> getTopIds();
 
         void printDescriptor(ExqDescriptor<T,U,V>* data);
         void printStuff();
 
         // The query processing interface is very simple:
-        void compareAndReplaceFarthest(ExqDescriptor<T,U,V>* data, uint64_t clusterid = -1);
+        void compareAndReplaceFarthest(ExqDescriptor<T,U,V>* data, uint32_t clusterid = -1);
 
         // Find distance to a descriptor
         double distance(ExqDescriptor<T,U,V>* data);
 
         void open();
 
-        uint64_t* nextClusterID();  // NOTE: This does not advance the scan!
-        uint64_t* next();
+        uint32_t* nextClusterID();  // NOTE: This does not advance the scan!
+        uint32_t* next();
 
         void close();
 
@@ -54,19 +54,19 @@ namespace exq {
         ExqFunctions<ExqDescriptor<T,U,V>>* _functions;
 
         // Query answer information
-        uint64_t _k;
-        uint64_t _neighbors;
+        uint32_t _k;
+        uint32_t _neighbors;
         vector<double> _distances;
-        vector<uint64_t> _descriptorIDs;
-        vector<uint64_t> _clusterIDs;
+        vector<uint32_t> _descriptorIDs;
+        vector<uint32_t> _clusterIDs;
 
         // Support for efficient computation
         // Index into the arrays where the farthest descriptor is found
-        uint64_t _farthest;
+        uint32_t _farthest;
 
 
         // Replace the farthest neighbor info with this info
-        void ReplaceFarthest(uint64_t id, uint64_t clusterid, double dist);
+        void ReplaceFarthest(uint32_t id, uint32_t clusterid, double dist);
 
         // Find the new farthest neigbor after replacing the old one
         void FindFarthest();
