+++ src/base/eCP/ECPNearestNeighbour.h
@@ -12,10 +12,10 @@ namespace exq {
     template<typename T, typename U, typename V>
     class ECPNearestNeighbour {
     public:
-        ECPNearestNeighbour(ExqDescriptor<T,U,V>* _query, uint64_t _k,
+        ECPNearestNeighbour(ExqDescriptor<T,U,V>* _query, uint32_t _k,
                             ExqFunctions<ExqDescriptor<T,U,V>>*& functions, int featureDimensions);
 
-        ECPNearestNeighbour(uint64_t _k,
+        ECPNearestNeighbour(uint32_t _k,
                             ExqFunctions<ExqDescriptor<T,U,V>>*& functions, int featureDimensions);
 
         ~ECPNearestNeighbour();
@@ -23,18 +23,18 @@ namespace exq {
         // Get info
         ExqDescriptor<T,U,V>* getQuery() { return query; }
 
-        uint64_t getK() { return k; }
+        uint32_t getK() { return k; }
 
         // Find distance to a descriptor
         double Distance(ExqDescriptor<T,U,V> *data);
 
         // The query processing interface is very simple:
-        void compareAndReplaceFarthest(ExqDescriptor<T,U,V>* data, uint64_t clusterId = -1);
+        void compareAndReplaceFarthest(ExqDescriptor<T,U,V>* data, uint32_t clusterId = -1);
 
         void open();
 
-        uint64_t* nextClusterID();  // NOTE: This does not advance the scan!
-        uint64_t* next();
+        uint32_t* nextClusterID();  // NOTE: This does not advance the scan!
+        uint32_t* next();
 
         double* getDistanceForDescriptor();
 
@@ -47,25 +47,25 @@ namespace exq {
         // Query information
         ExqDescriptor<T,U,V>* query;
         //int mod;
-        uint64_t numdims{};
+        uint32_t numdims{};
         ExqFunctions<ExqDescriptor<T,U,V>>* _functions;
         int _featureDimensions;
 
         // Query answer information
-        uint64_t k;
-        uint64_t neighbors;
+        uint32_t k;
+        uint32_t neighbors;
         double *distances;
-        uint64_t *descriptorIDs;
-        uint64_t *clusterIDs;
+        uint32_t *descriptorIDs;
+        uint32_t *clusterIDs;
 
         vector<double> extendedValueArr;
         vector<vector<double>> indexQueryArr;
         // Support for efficient computation
         // Index into the arrays where the farthest descriptor is found
-        uint64_t farthest;
+        uint32_t farthest;
 
         // Replace the farthest neighbor info with this info
-        void ReplaceFarthest(uint64_t id, uint64_t clusterId, double dist);
+        void ReplaceFarthest(uint32_t id, uint64_t clusterId, double dist);
 
         // Find the new farthest neigbor after replacing the old one
         void FindFarthest();
